test(size): keep the report in one file, drop the unit tests
The rendering was split into a helper only so unit tests could reach it.
Fold it back into the size harness and drop those tests — the report is
covered by running it, and this keeps the change to the template alone.
